Created stylesheet for carCard
